views/lighttable: Select the current displayed picture when zoom == 1.
fix #9123 and #9431

In filemanager mode and zoomable table with zoom == 1 (only one picture
displayed on screen), it is still possible to do several actions using
selected pictures tool (eg delete) on the current selected pictures.

The selection can be different from the picture displayed on screen.
To avoid confusing the user, when zoom == 1 the displayed picture becomes
automatically the currently selected picture so that the selected
pictures tool acts on the current displayed picture.

When unzooming, the previous selection is lost.