selection: Automatically select the full previewed picture.
This feature need to be discussed with darktable community as I don't
see a need to navigate through picture in full preview mode. Moreover I
think this is a bad idea to modify current selection when going to full
preview mode.

A better solution would be to disable any action on the full previewed
picture (including navigation with left/right/scroll).