lighttable: Update images_in_row variable.
The lib->images_in_row was only updated when scrolling with a modifier
key activated (always set to 0 in other case). It is needed by the
get_images_in_row function to export this value outside of the lighttable.c
file.

This commit updates lib->images_in_row on filemanager expose and
zoomable table expose.